Revert "Cherry pick PR #6257: Refactor Shell into cobalt directory. (#6279)"
This reverts commit 1f1cee03984b83ab530922ceca30175924331997.
