+++ cobalt/android/apk/app/src/org/chromium/content_shell/Util.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dev.cobalt.shell;
+package org.chromium.content_shell;
 
 import android.view.View;
 import android.view.ViewGroup;
